requirements.txt
gradio
openai-whisper
moviepy
torch
numpy
imageio[ffmpeg]
librosa
